fix(countdown): 适配v14 && type增加text类型
