+++ src/packages/countdown/countdown.harmony.css
@@ -3,54 +3,68 @@
   flex-direction: row;
   align-items: center;
   color: #ff0f23;
-  font-size: 10px;
+  font-size: 11px;
 }
 .nut-countdown-number-primary {
   display: flex;
   align-items: center;
   justify-content: center;
-  height: 14px;
+  height: 16px;
   font-weight: 400;
-  font-size: 10px;
+  font-size: 11px;
 }
 .nut-countdown-number {
   display: flex;
   align-items: center;
   justify-content: center;
-  height: 14px;
+  height: 16px;
   font-weight: 400;
-  font-size: 10px;
+  font-size: 11px;
+}
+.nut-countdown-number-text {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 16px;
+  font-weight: 400;
+  font-size: 11px;
 }
 .nut-countdown-unit {
   display: flex;
   align-items: center;
   justify-content: center;
-  height: 14px;
+  height: 16px;
   font-weight: 400;
-  font-size: 10px;
+  font-size: 11px;
 }
 .nut-countdown-number {
-  min-width: 20px;
-  padding: 0 1px;
+  min-width: 16px;
+  padding: 0 0;
   border-radius: 2px;
-  margin: 0 2px;
+  margin: 0 1px;
+  text-align: center;
 }
 .nut-countdown-number-primary {
-  min-width: 20px;
-  padding: 0 1px;
+  min-width: 16px;
+  padding: 0 0;
   border-radius: 2px;
-  margin: 0 2px;
+  margin: 0 1px;
+  text-align: center;
 }
 .nut-countdown-number {
   border: 1px solid #ffebf1;
-  background-color: transparent;
+  background-color: #ffffff;
   color: #ff0f23;
-  text-align: center;
 }
 .nut-countdown-number-primary {
   border: 1px solid #ff0f23;
   background-color: #ff0f23;
   color: #ffffff;
+}
+.nut-countdown-number-text {
+  border: 0;
+  background-color: transparent;
+  color: #ff0f23;
 }
 .nut-countdown-unit {
   color: #ff0f23;
