+++ src/packages/countdown/doc.taro.md
@@ -90,6 +90,7 @@ import { CountDown } from '@nutui/nutui-react-taro'
 
 | 属性 | 说明 | 类型 | 默认值 |
 | --- | --- | --- | --- |
+| type | 展示类型 | `default`\|`primary`\|`text` | `default` |
 | startTime | 开始时间 | `number` | `Date.now()` |
 | endTime | 结束时间 | `number` | `Date.now()` |
 | remainingTime | 剩余时间，单位是毫秒 | `number` | `0` |
@@ -134,12 +135,12 @@ import { CountDown } from '@nutui/nutui-react-taro'
 | --- | --- | --- |
 | \--nutui-countdown-display | 倒计时的布局方式 | `flex` |
 | \--nutui-countdown-color | 倒计时的文字颜色 | `$color-primary` |
-| \--nutui-countdown-font-size | 倒计时的字体大小 | `10px` |
+| \--nutui-countdown-font-size | 倒计时的字体大小 | `11px` |
 | \--nutui-countdown-font-weight | 倒计时的字体粗细 | `400` |
-| \--nutui-countdown-width | 倒计时的时间区域宽度 | `20px` |
-| \--nutui-countdown-height | 倒计时的时间区域高度 | `14px` |
-| \--nutui-countdown-number-padding | 倒计时的时间区域padding | `0 1px` |
-| \--nutui-countdown-number-margin | 倒计时的时间区域margin | `0 2px` |
+| \--nutui-countdown-width | 倒计时的时间区域宽度 | `16px` |
+| \--nutui-countdown-height | 倒计时的时间区域高度 | `16px` |
+| \--nutui-countdown-number-padding | 倒计时的时间区域padding | `0 0` |
+| \--nutui-countdown-number-margin | 倒计时的时间区域margin | `0 1px` |
 | \--countdown-number-border-radius | 倒计时的时间区域圆角大小 | `2px` |
 | \--nutui-countdown-color | 倒计时的时间区域的文字颜色 | `$color-primary-light-pressed` |
 | \--nutui-countdown-background-color | 倒计时的时间区域的背景颜色 | `$color-primary` |
