+++ src/packages/countdown/index.taro.ts
@@ -1,4 +1,4 @@
 import { CountDown } from './countdown.taro'
 
-export type { CountDownProps } from './countdown.taro'
+export type { CountDownProps, CountDownType } from './types'
 export default CountDown
