+++ src/packages/countdown/index.ts
@@ -1,4 +1,4 @@
 import { CountDown } from './countdown'
 
-export type { CountDownProps } from './countdown'
+export type { CountDownProps, CountDownType } from './types'
 export default CountDown
