+++ src/styles/variables.scss
@@ -775,19 +775,19 @@ $numberkeyboard-key-confirm-background-color: var(
 // countdown（✅）
 $countdown-display: var(--nutui-countdown-display, flex) !default;
 $countdown-color: var(--nutui-countdown-color, $color-primary) !default;
-$countdown-font-size: var(--nutui-countdown-font-size, 10px) !default;
-$coutdown-font-weight: var(--nutui-countdown-font-weight, 400) !default;
-$countdown-height: var(--nutui-countdown-height, 14px) !default;
-$countdown-width: var(--nutui-countdown-width, 20px) !default;
-$countdown-number-padding: var(
-  --nutui-countdown-number-padding,
-  0 1px
+$countdown-font-size: var(--nutui-countdown-font-size, $font-size-xs) !default;
+$coutdown-font-weight: var(
+  --nutui-countdown-font-weight,
+  $font-weight
 ) !default;
+$countdown-height: var(--nutui-countdown-height, 16px) !default;
+$countdown-width: var(--nutui-countdown-width, 16px) !default;
+$countdown-number-padding: var(--nutui-countdown-number-padding, 0 0) !default;
 $countdown-number-border-radius: var(
   --nutui-countdown-number-border-radius,
-  2px
+  $radius-xxs
 ) !default;
-$countdown-number-margin: var(--nutui-countdown-number-margin, 0 2px) !default;
+$countdown-number-margin: var(--nutui-countdown-number-margin, 0 1px) !default;
 $countdown-number-border-color: var(
   --nutui-countdown-number-border-color,
   $color-primary-light-pressed
@@ -798,7 +798,7 @@ $countdown-number-color: var(
 ) !default;
 $countdown-number-background-color: var(
   --nutui-countdown-number-background-color,
-  transparent
+  $color-background-overlay
 ) !default;
 $countdown-number-primary-border-color: var(
   --nutui-countdown-number-primary-border-color,
@@ -806,7 +806,7 @@ $countdown-number-primary-border-color: var(
 ) !default;
 $countdown-number-primary-color: var(
   --nutui-countdown-number-primary-color,
-  #ffffff
+  $color-primary-text
 ) !default;
 $countdown-number-primary-background-color: var(
   --nutui-countdown-number-primary-background-color,
